test: Generalize `processFunction` into global lib
